languages/python/ai_orchestrator/requirements.txt
fastapi
uvicorn
openai
python-multipart
boto3
stripe
dropbox
google-api-python-client
google-auth
google-auth-oauthlib
httpx
beautifulsoup4
